[관리자] - 상생사업관리 회원검색 오류 수정 - 1:1 문의 목록 키워드 검색 오류 수정 - 컨설팅사업관리 기술지도 신청정보 pdf 다운로드 영역 수정
[사용자]
- 안전설비구축 신청정보 SQ정보 하이픈 처리 수정

+++ service/src/main/java/com/kap/core/dto/cb/cba/CBATechGuidanceUpdateDTO.java
@@ -72,11 +72,11 @@ public class CBATechGuidanceUpdateDTO extends BaseDTO {
 
     // 사업진행 상세
     @Schema(title = "불량개선이전율")
-    private Integer fltyImpvmBfreRate;
+    private Double fltyImpvmBfreRate;
     @Schema(title = "불량개선이후률")
-    private Integer fltyImpvmAftrRate;
+    private Double fltyImpvmAftrRate;
     @Schema(title = "불량개선율")
-    private double fltyImpvmRate;
+    private Double fltyImpvmRate;
 
     @Schema(title = "초도방문자료")
     private Integer initVstFileSeq;
